backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.1
requests==2.32.3
python-dotenv==1.0.1
apscheduler==3.10.4
aiosqlite==0.20.0